Layout change required auth improvement
