+++ app/(auth)/register.tsx
